removal of unnecessary cast

+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/config/AmqpConfiguration.java
@@ -71,7 +71,7 @@ public class AmqpConfiguration {
         if (StringUtils.isNotBlank(properties.getPassword())) {
             factory.setPassword(properties.getPassword());
         }
-        factory.setCloseTimeout((long) properties.getCloseTimeout());
+        factory.setCloseTimeout(properties.getCloseTimeout());
 
         return factory;
     }
